added toasts, navigate to profile when save photo, change instructions list, and condition to showing original photo in modal
