+++ package.json
@@ -33,6 +33,7 @@
     "react-scripts": "5.0.1",
     "react-select": "^5.10.1",
     "react-simple-typewriter": "^5.0.1",
+    "react-toastify": "^11.0.5",
     "redux-persist": "^6.0.0",
     "sass": "^1.86.3",
     "styled-components": "^6.1.17",
