+++ public/locales/en/translation.json
@@ -137,8 +137,8 @@
     "attentionText": "*You have 10 free generations",
     "mainWindowDesc": "Display or change your exterior design for a limited time. Just upload a photo or sketch and watch the magic in action.",
     "mainWindowAdvance_0": "Upload an image of your design",
-    "mainWindowAdvance_1": "Select the style and type of room",
-    "mainWindowAdvance_2": "Write a prompt, describe what you want to achieve. Describe colors or other details",
+    "mainWindowAdvance_2": "Select the style and type of room",
+    "mainWindowAdvance_1": "Write a prompt, describe what you want to achieve. Describe colors or other details",
     "btnUpgrade": "Upgrade",
     "fileClick": "Click to upload",
     "fileDrag": "or drag and drop",
