+++ public/locales/ua/translation.json
@@ -137,8 +137,8 @@
     "attentionText": "*у вас є 10 безкоштовних генерацій",
     "mainWindowDesc": "Рендеріть або змінюйте свій дизайн екстер’єру за лічені секунди. Просто завантажте фотографію чи ескіз і подивіться на магію в дії.",
     "mainWindowAdvance_0": "Завантажте зображення свого дизайну",
-    "mainWindowAdvance_1": "Виберіть стиль та вид кімнати",
-    "mainWindowAdvance_2": "Напишіть prompt, опишіть який результат хочете отримати. Опишіть кольори чи інші деталі",
+    "mainWindowAdvance_2": "Виберіть стиль та вид кімнати",
+    "mainWindowAdvance_1": "Напишіть prompt, опишіть який результат хочете отримати. Опишіть кольори чи інші деталі",
     "btnUpgrade": "Оновити",
     "fileClick": "Натисніть",
     "fileDrag": "або перенесіть файл",
@@ -189,7 +189,7 @@
     "folderName": "Назва папки",
     "folderInputPlaceholder": "Наприклад \"Кухня\"",
     "folderExist": "Папка з такою назвою вже існує",
-    "createdFolder": "Папку успішно створена"
+    "createdFolder": "Папку успішно створено"
   },
   "profile": {
     "yourFolders": "Ваші папки",
