+++ src/_components/AddCollectionModalFoldersList/AddCollectionModalFolersList.styled.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const AddCollectionModalFoldersListStyles = styled.div`
+  h1 {
+    margin-bottom: 40px;
+    font-size: 20px;
+  }
+
+  ul {
+    display: flex;
+    flex-direction: column;
+    row-gap: 8px;
+    margin-bottom: 32px;
+  }
+`;
