+++ src/_components/CreatingFolderModal/CreatingFolderModal.jsx
@@ -1,19 +1,24 @@
 import { CreatingFolderModalStyles } from "./CreatingFolderModalStyles.styled";
 import { useTranslation } from "react-i18next";
 import CloseModalButton from "../CloseModalButton/CloseModalButton";
-import { useSelector } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import { getGenerationFolders } from "../../redux/generationFolders/generationFoldersSelectors";
 import { useState } from "react";
+import { createFolder } from "../../redux/generationFolders/generationFoldersSlice";
+import { toastSuccess } from "../../assets/functions/toastNotification";
+import { useNavigate } from "react-router-dom";
 
 const CreatingFolderModal = ({
   newFolderName,
   setNewFolderName,
   toggleModal,
-  creatingFolder,
+  result,
 }) => {
   const { t } = useTranslation();
   const folders = useSelector(getGenerationFolders);
   const [errorMessage, setErrorMessage] = useState("");
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
 
   function checkFolderNames(value) {
     if (folders.some(folder => folder.title === value)) {
@@ -25,18 +30,14 @@ const CreatingFolderModal = ({
 
   function handleChangeName(e) {
     setNewFolderName(e.target.value);
-    checkFolderNames(e.target.value);
+    checkFolderNames(e.target.value.trim());
   }
 
   function handleSubmit(e) {
     e.preventDefault();
-    creatingFolder();
-
-    setTimeout(() => {
-      toggleModal();
-      // Перенаправляем пользователя в личный кабинет
-      // ........
-    }, 3000);
+    dispatch(createFolder({ folderTitle: newFolderName, photo: result }));
+    toastSuccess(t("modal.createdFolder"));
+    navigate("/profile/main");
   }
 
   const placeholderValue = t("modal.folderInputPlaceholder");
