+++ src/pages/GeneratePage.jsx
@@ -4,7 +4,7 @@ import GeneratingAnswer from "../_components/GeneratingAnswer/GeneratingAnswer";
 import { GeneratingPageStyles } from "./styles/GeneratingPage.styled";
 import MainContainer from "../_components/MainContainer/MainContainer";
 import GeneralModal from "../_components/GeneralModal/GeneralModal";
-import AddCollectionModal from "../_components/AddCollectionModal/AddCollectionModal";
+import AddCollectionModalContainer from "../_components/AddCollectionModalContainer/AddCollectionModalContainer";
 import { addCollectionStyles } from "../_components/AddCollectionModal/addCollectionStylesProps";
 
 const GeneratePage = () => {
@@ -43,7 +43,7 @@ const GeneratePage = () => {
           styleSizes={addCollectionStyles}
           toggleModal={toggleModal}
         >
-          <AddCollectionModal result={result} />
+          <AddCollectionModalContainer result={result} />
         </GeneralModal>
       )}
     </>
